refactor: split route logic into domain services
Extract user, group, and request service modules so route handlers delegate database lookup, mutation, serialization, and commit handling to focused service functions.

Move asset serializers into asset_service, add generic group asset listing for jobs and structures, and keep group-member visibility filtering in one shared path.

Replace sender/receiver-specific request getters with get_request_or_404 plus explicit sender/receiver authorization helpers that preserve 404 responses for unauthorized request access.

Reject group_admin role updates that omit group_id, remove the duplicated query_helpers module, and update tests around admin role changes, request authorization, asset serialization, and shared service lookups.

Validation: PYTHONPATH=. venv/bin/python -m pytest -p no:capture -q
