+++ tests/test_user_group_services.py
@@ -3,36 +3,27 @@ import uuid
 import pytest
 from fastapi import HTTPException
 
-from conftest import make_auth0_payload
-from query_helpers import (
-    get_current_user_or_404,
-    get_group_or_404,
-)
+from group_service import get_group_or_404
+from user_service import get_user_or_404
 
 
-class TestCurrentUserQueries:
-    def test_get_current_user_returns_persisted_user(self, db, user_factory):
+class TestUserService:
+    def test_get_user_returns_persisted_user(self, db, user_factory):
         user = user_factory(user_sub="auth0|current")
 
-        result = get_current_user_or_404(
-            db,
-            make_auth0_payload(user.user_sub),
-        )
+        result = get_user_or_404(db, user.user_sub)
 
         assert result is user
 
-    def test_get_current_user_returns_404_when_missing(self, db):
+    def test_get_user_returns_404_when_missing(self, db):
         with pytest.raises(HTTPException) as error:
-            get_current_user_or_404(
-                db,
-                make_auth0_payload("auth0|missing"),
-            )
+            get_user_or_404(db, "auth0|missing")
 
         assert error.value.status_code == 404
         assert error.value.detail == "User not found"
 
 
-class TestEntityQueries:
+class TestGroupService:
     def test_get_group_returns_persisted_group(self, db, group_factory):
         group = group_factory()
